Fix a 64-bitism, use size_t instead of u_int for 4th arg to copyinstr.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@65722 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/miscfs/kernfs/kernfs_vfsops.c
@@ -74,7 +74,7 @@ kernfs_mount(mp, path, data, ndp, p)
 	struct proc *p;
 {
 	int error = 0;
-	u_int size;
+	size_t size;
 	struct kernfs_mount *fmp;
 	struct vnode *rvp;
